package rename tv -> automotive
