+++ app-automotive/proguard-rules.pro
