+++ app-automotive/src/main/AndroidManifest.xml
