+++ app-automotive/src/main/res/mipmap-mdpi/ic_launcher_foreground.webp
