+++ app-automotive/src/main/res/values/colors.xml
