+++ app-automotive/src/main/res/values/strings.xml
