+++ app-automotive/src/main/res/xml/automotive_app_desc.xml
